Updated: Render Layer handle transparency (WIP).
