+++ libui/includes/privates/ui_shared.h
@@ -6,7 +6,7 @@
 /*   By: alerandy <alerandy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/26 18:03:13 by alerandy          #+#    #+#             */
-/*   Updated: 2019/06/06 11:22:05 by alerandy         ###   ########.fr       */
+/*   Updated: 2019/07/23 13:00:11 by alerandy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,6 @@
 
 char	*get_file_name(char *path);
 void	close_fd(int fd, char *error);
+void	convert_color_lines(unsigned *dst, unsigned *src, unsigned width);
 
 #endif
