+++ libui/src/ui_layers/ui_layer_into_layer.c
@@ -6,41 +6,13 @@
 /*   By: alerandy <alerandy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/07 12:52:23 by alerandy          #+#    #+#             */
-/*   Updated: 2019/06/07 15:00:05 by alerandy         ###   ########.fr       */
+/*   Updated: 2019/07/23 13:01:04 by alerandy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libui_layers.h"
 #include "bmp_parser.h"
-
-static inline unsigned	merge_pixel(unsigned dst, unsigned src)
-{
-	double			alpha;
-	t_bmp_32		merge;
-	t_bmp_32		source;
-
-	merge = hex_to_bit32_pixel(dst);
-	source = hex_to_bit32_pixel(src);
-	alpha = (255 - source.a) / 255.;
-	merge.r = merge.r * alpha + source.r * (1 - alpha);
-	merge.g = merge.g * alpha + source.g * (1 - alpha);
-	merge.b = merge.b * alpha + source.b * (1 - alpha);
-	dst = bit32_pixel_to_hex(merge);
-	return (dst);
-}
-
-static inline void		convert_color_lines(unsigned *dst, unsigned *src, \
-												unsigned width)
-{
-	unsigned	i;
-
-	i = 0;
-	while (i < width)
-	{
-		*(dst + i) = merge_pixel(*(dst + i), *(src + i));
-		++i;
-	}
-}
+#include "ui_shared.h"
 
 void					ui_layer_into_layer(t_ui_layer *dst, t_ui_layer *src)
 {
