+++ libui/src/ui_shared/make.dep
@@ -1,1 +1,1 @@
-SRCS += get_file_name.c ui_close_fd.c 
+SRCS += get_file_name.c ui_close_fd.c ui_pixscpy.c
